Selection mode of photo collection view. Change navigation bar, hide tab bar and show toolbar.
It is not allowed to add a tool bar to a UICollectionViewController. When I use toolbar of the parent UINavigationController, it appears above the tab bar even the tab bar is hidden.
I believe this is because UINavigationController is embedded in a UITabBarController, but I don’t know how to fix this. So I use a UIViewController with a toolbar and a collection view inside it instead of using a UICollectionViewController.